+++ src/game/player/elements/menu/Menu.tsx
@@ -6,6 +6,7 @@ import {MenuElementProps} from "@player/elements/menu/type";
 import Isolated from "@player/lib/isolated";
 import {useGame} from "@player/provider/game-state";
 import Sentence from "@player/elements/say/Sentence";
+import Inspect from "@player/lib/Inspect";
 
 export default function Menu(
     {
@@ -25,23 +26,25 @@ export default function Menu(
     return (
         <>
             <Isolated className={"absolute"}>
-                <div className="absolute flex flex-col items-center justify-center min-w-full w-full h-full">
-                    {prompt && <Say
-                        state={state}
-                        action={{sentence: prompt, character: null}}
-                        useTypeEffect={false}
-                        className="z-10"
-                    />}
-                </div>
+                {prompt && <Say
+                    state={state}
+                    action={{sentence: prompt, character: null}}
+                    useTypeEffect={false}
+                    className="z-10"
+                />}
             </Isolated>
-            <div className={clsx(
-                "absolute flex flex-col items-center justify-center min-w-full w-full h-full",
-                game.config.elementStyles.menu.containerClassName
-            )}>
+            <Inspect.Div
+                tag={"menu.containerClassName"}
+                className={clsx(
+                    "absolute flex flex-col items-center justify-center min-w-full w-full h-full",
+                    game.config.elementStyles.menu.containerClassName
+                )}
+            >
                 <div className="p-4 rounded-lg w-full z-20">
                     <div className="flex flex-col items-center mt-4 w-full">
                         {choices.map((choice, i) => (
-                            <button
+                            <Inspect.Button
+                                tag={"menu.choiceButtonClassName." + i}
                                 key={i}
                                 className={clsx(
                                     "bg-white text-black p-2 mt-2 w-1/2",
@@ -55,11 +58,11 @@ export default function Menu(
                                     useTypeEffect={false}
                                     className={clsx(game.config.elementStyles.menu.choiceButtonTextClassName)}
                                 />
-                            </button>
+                            </Inspect.Button>
                         ))}
                     </div>
                 </div>
-            </div>
+            </Inspect.Div>
         </>
     );
 };